% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_prewhiten.R
\name{ar_prewhiten}
\alias{ar_prewhiten}
\title{Autoregressive Prewhitening}
\usage{
ar_prewhiten(x, return = "both", ...)
}
\arguments{
\item{x}{univariate or multivariate time series}

\item{return}{character string, determines whether to return model estimates ("est"), the ar model ("model"), or both ("both")}
}
\value{
model estimates, ar model object, or a list containing both
}
\description{
Autoregressive Prewhitening
}
\details{
Function modified from ar.prewhiten in the dplR package by Andy Bunn.
}
\examples{
# simulate autocorrelated timeseries
ar.sim<-arima.sim(model=list(ar=c(.9,-.2)),n=100)

#prewhiten
white <- ar_prewhiten(ar.sim, return = "both" )

plot(ar.sim)
lines(white[[1]], col = "red")

}
