% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg_func.R
\name{calc_PCs}
\alias{calc_PCs}
\title{calculate PCs}
\usage{
calc_PCs(periods_df, PCA_chrons, pc.calc, nest_yrs, calib, full)
}
\arguments{
\item{periods_df}{dataframe of nests}

\item{PCA_chrons}{dataframe containing chronologies}

\item{period}{period of time over which to calculate PCs, calibration or nest_yrs}
}
\value{

}
\description{
calculate PCs
}
