% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_clim.R
\name{eval_clim}
\alias{eval_clim}
\title{Evaluate climate/tree ring relationships and select chronologies that meet threshold}
\usage{
eval_clim(
  crns,
  lead = 1,
  lag = NULL,
  prewhiten_crn = TRUE,
  climate,
  mos = 5:8,
  method = "mean",
  calib,
  valid,
  cor_window = "calib",
  type = "pearson",
  alternative = "two.sided",
  r = 0.25,
  alpha = 0.9,
  print = TRUE,
  out_fmt = "R",
  out_dir = "PCregOutput/"
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Evaluate climate/tree ring relationships and select chronologies that meet threshold
}
