% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cor.R
\name{filter_cor}
\alias{filter_cor}
\title{Filter chronologies based on their correlation with climate data}
\usage{
filter_cor(
  crns,
  lead = 1,
  lag = NULL,
  clim,
  cor_window,
  type = "pearson",
  alternative = "two.sided",
  r = 0.25,
  alpha = 0.9,
  prewhiten_crn = TRUE,
  prewhiten_clim = TRUE,
  calib = calib,
  valid = valid,
  full = full
)
}
\arguments{
\item{crns}{dataframe, such as returned by the load_crns function, where there is a column containing years and one column for each chronology}

\item{lead}{single value indicating the maximum number of years to shift chronology data forward. 0 indicates no lead, 2 will evaluate 0, 1, and 2 year leads, etc. Defaults to 1.}

\item{clim}{dataframe, such as returned by the get_clim function, containing columns for year and either: individual months climate data (may be more than one), a sum of >1 months climate data, or the average of >1 months data}

\item{alternative}{indicates the alternative hypothesiss and must be one of "two.sided", "greater", or "less".}

\item{r}{minimum correlation coefficient needed to retain chronology. Defaults to 0.25.}

\item{alpha}{minimum alpha value for rejection of null hypothesis. Defaults to 0.90}

\item{full}{}
}
\value{

}
\description{
Filter chronologies based on their correlation with climate data
}
