% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spat.R
\name{filter_foot}
\alias{filter_foot}
\title{Spatial Filtering of Points Based on Climate Footprint}
\usage{
filter_foot(
  x,
  footprint,
  r = 0.3,
  cent.lat,
  cent.lon,
  radius = 500,
  plot = TRUE
)
}
\arguments{
\item{footprint}{Raster or SpatialPolygonDataFrame object}

\item{r}{minimum correlation coefficient for spatial selection}

\item{plot}{}

\item{data}{dataframe containing columns labeled ID, lat, and lon}
}
\value{
character vector of chronology names that fall within the area where correlation >= r
}
\description{
Spatial Filtering of Points Based on Climate Footprint
}
\examples{
# Climate footprint .nc file is a netCDF file downloaded from KNMI climate explorer with correlations between PRISM precipitation (1895-present) and streamflow at Gallinas Creek near Montezuma, NM.

#' footprint <- raster::raster("inst/extdata/gallinas_cf.nc")
select_crns_fp <- filter_foot(data = metadata, footprint = footprint, r = 0.5)

}
