% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spat.R
\name{filter_rad}
\alias{filter_rad}
\title{Spatial Filtering of Points Based on Radius From a Point}
\usage{
filter_rad(
  x,
  cent.lat,
  cent.lon,
  radius = 150,
  climate = NULL,
  plot = TRUE,
  buff = 5
)
}
\arguments{
\item{cent.lat}{latitude of center point}

\item{cent.lon}{longitude of center point}

\item{radius}{distance from center in km, default 150}

\item{data}{dataframe containing columns labeled ID, lat, and lon}
}
\value{
character vector of chronology names within the selected radius
}
\description{
Spatial Filtering of Points Based on Radius From a Point
}
\examples{
# metadata returned from parse_json function, must be a dataframe or matrix containing columns lat, lon, and ID
select_crns_rad <- filter_rad(data = metadata, cent.lat = 35.65, cent.lon = -105.32, radius = 150)

}
