% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spat.R
\name{filter_spat}
\alias{filter_spat}
\title{Spatial Filtering of Tree Ring Data}
\usage{
filter_spat(ID, lat, lon, center, radius, climate = NULL, type = "radius")
}
\arguments{
\item{ID}{vector of character strings containing chronology IDs}

\item{lat}{numeric vector of latitudes for each chronology}

\item{lon}{numeric vector of longitudes for each chronology}

\item{center}{center point for calculating radius (radius type)}

\item{radius}{distance from center in km, default 150 (radius type)}

\item{climate}{dataframe with climate variables for spatial correlation (footprint type), defaults to NULL}

\item{type}{"radius" or "footprint", defaults to radius}
}
\value{

}
\description{
Spatial Filtering of Tree Ring Data
}
