% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clim.R
\name{load_clim}
\alias{load_clim}
\title{Get months of interest for climate analysis, sum or average them}
\usage{
load_clim(clim, mos, method = "mean", prewhiten_clim = TRUE, full = full)
}
\arguments{
\item{mos}{months of interest}
}
\value{
dataframe of monthly data
}
\description{
Get months of interest for climate analysis, sum or average them
}
