% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg_func.R
\name{mod_df}
\alias{mod_df}
\title{Create dataframe containing clim and select PCs for linear model}
\usage{
mod_df(data = PC_vals, eig = eigval_small, nest_yrs, calib)
}
\arguments{
\item{data}{PCA values}

\item{eig}{selected eigenvalues and PCs}
}
\value{

}
\description{
Create dataframe containing clim and select PCs for linear model
}
