% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_tbl.R
\name{nest_tbl}
\alias{nest_tbl}
\title{Creates dataframe of chronology names, start years, and end years from chronology dataframe}
\usage{
nest_tbl(select_crns_cor)
}
\arguments{
\item{select_crns_cor}{dataframe of chronology values where the first column is years, and each subsequent column is a chronologies}
}
\value{
dataframe of chronology names, start years, and end years
}
\description{
Creates dataframe of chronology names, start years, and end years from chronology dataframe
}
