% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrecon.R
\docType{package}
\name{pcrecon}
\alias{pcrecon}
\title{pcrecon: A package for conducting principle component regression on tree ring data for climate reconstruction}
\description{
The pcrecon package provides three categories of important functions:
foo, bar and baz.
}
\section{pcrecon functions}{

The pcreg functions ...
}

