% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spat.R
\name{radius}
\alias{radius}
\title{Radius Polygon Around Point}
\usage{
radius(cent_lat, cent_lon, radius)
}
\arguments{
\item{radius}{}
}
\value{

}
\description{
Radius Polygon Around Point
}
\examples{

}
