% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crn_head.R
\name{read_crn_head}
\alias{read_crn_head}
\title{Parse chronology file headers}
\usage{
read_crn_head(fname, err = "warn", logfile = "read_crns.log")
}
\arguments{
\item{fname}{character string of full path to file and including file ending}

\item{err}{character string indicating type of error handling where `stop` will throw and error, `warn` will print warning to the screen, and `log` will pass the test for use in a log file. Defaults to `log`.}

\item{logfile}{character string indicating path to the log file for errors associated with reading in chronologies. The path is the same as `dir` if not otherwise specified.}
}
\value{
dataframe of fname, site_id, site_name, sp_code (species code), type_c (chronology type), type_m (measurement type), elev, lat_lon, years
}
\description{
Parse chronology file headers
}
\examples{
 read_crn_head(system.file("extdata/crns/nm575.crn", package = "pcreg"))
}
