% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save data from PCreg objects}
\usage{
save_data(data, out_fmt = out_fmt, out_dir = out_dir)
}
\arguments{
\item{data}{list containing objects to sabe}
}
\value{

}
\description{
Save data from PCreg objects
}
