% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_var.R
\name{scale_var}
\alias{scale_var}
\title{Scale variance to calibration or full observed period}
\usage{
scale_var(
  x = recon_nest,
  y = clim_full,
  scale.var = scale.var,
  calib,
  step_mod
)
}
\arguments{
\item{calib}{}
}
\value{

}
\description{
Scale variance to calibration or full observed period
}
