% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_PCs.R
\name{select_PCs}
\alias{select_PCs}
\title{select PCs}
\usage{
select_PCs(data = PCA, type = select.pc, cum.perc = NULL, m = NULL)
}
\arguments{
\item{data}{list object from calc_PCs}

\item{type}{eigval1, firstm, cum.perc, mean, and all}

\item{m}{}
}
\value{

}
\description{
select PCs
}
