% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg_func.R
\name{sum_mod}
\alias{sum_mod}
\title{model stats for table}
\usage{
sum_mod(data = step_mod, i, nest_yrs)
}
\arguments{
\item{data}{AIC selected linear model object}
}
\value{

}
\description{
model stats for table
}
