% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cor.R
\name{window_filter}
\alias{window_filter}
\title{Select windows for screening by correlation}
\usage{
window_filter(crns, clim, cor_window, calib, valid, full)
}
\arguments{
\item{clim}{}
}
\value{

}
\description{
Select windows for screening by correlation
}
