% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_prewhiten.R
\name{ar_prewhiten}
\alias{ar_prewhiten}
\title{Autoregressive Prewhitening}
\usage{
ar_prewhiten(x, return = "both", ...)
}
\arguments{
\item{x}{univariate or multivariate time series}

\item{return}{character string, determines whether to return model residuals ("resid"), the ar model ("model"), or both ("both")}
}
\value{

}
\description{
Autoregressive Prewhitening
}
\details{
Function modified from ar.prewhiten in the dplR package
}
