% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_prewhiten.R
\name{ar_prewhiten}
\alias{ar_prewhiten}
\title{Autoregressive Prewhitening}
\usage{
ar_prewhiten(x, return = "both", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Autoregressive Prewhitening
}
