% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PCs.R
\name{calc_PCs}
\alias{calc_PCs}
\title{calculate PCs}
\usage{
calc_PCs(period_df, PCA_chrons, pc_calc, nest_yrs, calib, full, valid)
}
\arguments{
\item{period_df}{}

\item{PCA_chrons}{dataframe containing chronologies}

\item{pc_calc}{period over which to calculate principal components ("calib", "valid", or "full")}

\item{nest_yrs}{years of the current nest being analyzed}

\item{calib}{calibration period years}

\item{full}{calibration and validation period}

\item{valid}{validation period years}
}
\value{

}
\description{
calculate PCs
}
\examples{



}
