% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cor.R
\name{crns_table}
\alias{crns_table}
\title{Fill table of chronologies and lead chronologies based on correlation output}
\usage{
crns_table(cors_table_small)
}
\arguments{
\item{cors_table_small}{dataframe of filtered chronology names, leads, correlation, and p values.}
}
\value{
dataframe of chronologies, including lead chronologies, that met threshold correlation values for analysis
}
\description{
Fill table of chronologies and lead chronologies based on correlation output
}
\examples{

}
