% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_clim.R
\name{eval_clim}
\alias{eval_clim}
\title{Evaluate climate/tree ring relationships and select chronologies that meet threshold}
\usage{
eval_clim(
  crns,
  lag = 1,
  prewhiten.crn = TRUE,
  climate,
  mos = 5:8,
  method = "mean",
  prewhiten.clim = TRUE,
  calib,
  valid,
  cor.window = "calib",
  type = "pearson",
  alternative = "two.sided",
  r = 0.25,
  alpha = 0.9,
  print.out = TRUE,
  save.out = "R",
  dir = "PCregOutput/"
)
}
\arguments{
\item{dir}{}
}
\value{

}
\description{
Evaluate climate/tree ring relationships and select chronologies that meet threshold
}
