% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spat.R
\name{filter_foot}
\alias{filter_foot}
\title{Spatial Filtering of Points Based on Climate Footprint}
\usage{
filter_foot(data, footprint, r, alpha)
}
\arguments{
\item{data}{dataframe containing columns labeled ID, lat, and lon}

\item{footprint}{Raster stack object}

\item{r}{minimum correlation coefficient for spatial selection}

\item{alpha}{alpha value , default null because you do this when making the footprint in KNMI}
}
\value{

}
\description{
Spatial Filtering of Points Based on Climate Footprint
}
\examples{

}
