% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_crns.R
\name{load_crns}
\alias{load_crns}
\title{Title}
\usage{

}
\arguments{
\item{dir}{path to directory containing chronologies as character string}

\item{crns}{vector of character strings containing the names of the chronologies wanting to read into R}

\item{logfile}{character string indicating path to the log file for errors associated with reading in chronologies. The path is the same as `dir` if not otherwise specified.}
}
\value{
dataframe with a year column and columns for each chronology containing values
}
\description{

}
\details{
coming soon
}
\examples{
\dontrun{
df <- load_crns(dir = system.file("extdata/crns", package = "pcreg", mustWork = TRUE), crns = list.files(system.file("extdata/crns", package = "pcreg", mustWork = TRUE)))
}
}
