% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_json}
\alias{parse_json}
\title{Parse metadata json files from ITRDB}
\usage{
parse_json(dir)
}
\arguments{
\item{dir}{file directory where json files are saved}
}
\value{
dataframe of ID, lat, lon, elevation
}
\description{
Parse metadata json files from ITRDB
}
\examples{
\dontrun{

}
}
