% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg.R, R/pcreg_func.R
\docType{package}
\name{pcreg}
\alias{pcreg}
\title{pcreg: A package for conducting principle component regression on tree ring data for climate reconstruction}
\usage{
pcreg(
  data,
  pc_calc = "calib",
  select_pc = "eigenvalue1",
  cum_perc = NULL,
  m = NULL,
  scale_var = "calib",
  weight = NULL,
  plot = TRUE,
  out_fmt = "csv"
)
}
\arguments{
\item{out_fmt}{}
}
\value{

}
\description{
The pcreg package provides three categories of important functions:
foo, bar and baz.
}
\section{pcreg functions}{

The pcreg functions ...
}

\examples{

}
