% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg_func.R
\name{pcreg}
\alias{pcreg}
\title{PCreg}
\usage{
pcreg(
  data,
  pc_calc = "calib",
  select_pc = "eigenvalue1",
  cum_perc = NULL,
  m = NULL,
  scale_var = "calib",
  weight = NULL,
  plot = TRUE,
  out_fmt = "csv"
)
}
\arguments{
\item{out_fmt}{}
}
\value{

}
\description{
PCreg
}
\examples{

}
