% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg.R, R/pcreg_func.R
\docType{package}
\name{pcreg}
\alias{pcreg}
\title{pcreg: A package for conducting principle component regression on tree ring data for climate reconstruction}
\usage{
pcreg(
  crns,
  lead = 1,
  prewhiten.crn = TRUE,
  climate,
  mos = 5:8,
  method = "mean",
  prewhiten.clim = FALSE,
  cor.window = "calib",
  type = "pearson",
  alternative = "two.sided",
  r = 0.25,
  alpha = 0.9,
  calib,
  valid,
  pc.calc = "calib",
  select.pc = "eigenvalue1",
  scale.var = "calib",
  weight = NULL
)
}
\arguments{
\item{weight}{}
}
\value{

}
\description{
The pcreg package provides three categories of important functions:
foo, bar and baz.
}
\section{pcreg functions}{

The pcreg functions ...
}

