% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg.R, R/pcreg_func.R
\docType{package}
\name{pcreg}
\alias{pcreg}
\title{pcreg: A package for conducting principle component regression on tree ring data for climate reconstruction}
\usage{
pcreg(clim, crns, nests, calib, valid, pc.calc, scale.var, weight)
}
\arguments{
\item{weight}{}
}
\value{

}
\description{
The pcreg package provides three categories of important functions:
foo, bar and baz.
}
\section{pcreg functions}{

The pcreg functions ...
}

