% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcreg.R
\docType{package}
\name{pcreg}
\alias{pcreg}
\title{pcreg: A package for conducting principle component regression on tree ring data for climate reconstruction}
\description{
The pcreg package provides three categories of important functions:
foo, bar and baz.
}
\section{pcreg functions}{

The pcreg functions ...
}

