% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_stats.R
\name{perf_stats}
\alias{perf_stats}
\title{Performance statistics of calibration and validation date from reconstruction}
\usage{
perf_stats(
  valid_est,
  calib_est,
  observed,
  valid_yrs = valid,
  calib_yrs = calib,
  mod_id = "m"
)
}
\arguments{
\item{valid_est}{dataframe of years and fit estimates from validation period}

\item{observed}{dataframe of years and observed values}

\item{valid_yrs}{integer vector of years for validation}

\item{calib_yrs}{integer vector of years used in calibration}

\item{mod_id}{character string identifying the model}
}
\value{
R2, Pearson_R2, RE, CE for validation data
}
\description{
Performance statistics of calibration and validation date from reconstruction
}
\details{
coming soon
}
\examples{
\dontrun{
"coming soon"
}
}
