% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_crn_head.R
\name{read_crn_head}
\alias{read_crn_head}
\title{Parse chronology file headers}
\usage{
read_crn_head(fname)
}
\arguments{
\item{fname}{character string of full path to file and including file ending}
}
\value{
dataframe of site_id, site_name, species_code, type_c, type_m, elev, lat_lon, years, first, last, fname
}
\description{
Parse chronology file headers
}
\examples{
 read_crn_head(system.file("extdata/crns/nm575.crn", package = "pcreg"))
}
