% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save data from PCreg objects}
\usage{
save_data(data, save.out = "R", dir)
}
\arguments{
\item{data}{list containing objects to sabe}

\item{save.out}{how to save the object: comma- delimited, text, or R data. Must be one of "csv", "table", or "R".}

\item{dir}{directory}
}
\value{

}
\description{
Save data from PCreg objects
}
